Extend GitPack database to query almost any URL
Grep processing during a database query has been replaces by a custom
solution processing each line. Yeah, it is a shell text processing but
I was not able to come up with a better solution - keep in mind that
URL must be matched as a fixed string and it must be matched only
against the URL field of a database record.